refactor: transaction detail
