+++ packages/helix/hooks/index.ts
@@ -1,1 +1,2 @@
 export * from './statistic';
+export * from './updatableRecord';
